Update SuCheck.cpp
I've had to consider several implications in the subBoard check that were a little more complicated than I had previously anticipated. I think that incrementing to each consecutive subBoard in the game board should hopefully be on the right track. The actual check performed per each subBoard was a little more difficult for me and required some more thought, so I'm less confident in that, but hopefully it is on the right track. Whereas you used the "i + 1" method in the row and column check functions, I allowed the subBoard check function to test at the i value, and compensated for this in the final if statement, where it accounts for the subBoard having exactly 1 "duplicate" value, which is in reality the subBoard counting the value for that initial time. Hopefully in practice this will work, but I won't be shocked if there is something I did not properly account for here. Let me know if you have any questions or insights, otherwise maybe we'll catch it once we start testing the class.